refactor making module with functions used across scripts
